requirements.txt
google-api-python-client==1.6.2
httplib2==0.10.3
oauth2client==4.1.0
protobuf==3.3.0
pyasn1==0.2.3
pyasn1-modules==0.0.8
python-dateutil==2.6.0
rsa==3.4.2
six==1.10.0
SQLAlchemy==1.1.9
uritemplate==3.0.0
wiringpi==2.44.0
